fix multiple traverse issue
