+++ Makefile
@@ -22,8 +22,8 @@ test_extract_developer_comments:
 test_extract_developer_comments_by_tag:
 	$(MOCHA_CMD) ./tests/functional/test_extract_developer_comments_by_tag.js
 
-test_extract_ngettext_with_multiple_presets:
-	$(MOCHA_CMD) ./tests/functional/test_extract_ngettext_with_multiple_presets.js
+test_extract_ngettext:
+	$(MOCHA_CMD) ./tests/functional/test_extract_ngettext.js
 
 test_resolve_gettext:
 	$(MOCHA_CMD) ./tests/functional/test_resolve_gettext.js
@@ -119,7 +119,7 @@ test_fun: test_extract_gettext_with_formatting
 test_fun: test_extract_filename
 test_fun: test_extract_developer_comments
 test_fun: test_extract_developer_comments_by_tag
-test_fun: test_extract_ngettext_with_multiple_presets
+test_fun: test_extract_ngettext
 test_fun: test_resolve_gettext
 test_fun: test_resolve_gettext_default
 test_fun: test_resolve_default
